+++ packages/core-cairo/src/erc20.ts
@@ -11,6 +11,8 @@ import { contractDefaults as commonDefaults } from './common-options';
 import { printContract } from './print';
 import { externalTrait } from './external-trait';
 import { toByteArray, toFelt252 } from './utils/convert-strings';
+import { addVotesComponent } from './common-components';
+
 
 export const defaults: Required<ERC20Options> = {
   name: 'MyToken',
@@ -53,7 +55,7 @@ function withDefaults(opts: ERC20Options): Required<ERC20Options> {
     mintable: opts.mintable ?? defaults.mintable,
     votes: opts.votes ?? defaults.votes,
     appName: opts.appName ?? defaults.appName,
-    appVersion: opts.appVersion ?? defaults.appVersion,
+    appVersion: opts.appVersion ?? defaults.appVersion
   };
 }
 
@@ -95,7 +97,8 @@ export function buildERC20(opts: ERC20Options): Contract {
 }
 
 function addHooks(c: ContractBuilder, allOpts: Required<ERC20Options>) {
-  if (allOpts.votes || allOpts.pausable) {
+  const usesCustomHooks = allOpts.pausable || allOpts.votes;
+  if (usesCustomHooks) {
     const hooksTrait = {
       name: 'ERC20HooksImpl',
       of: 'ERC20Component::ERC20HooksTrait<ContractState>',
@@ -104,31 +107,20 @@ function addHooks(c: ContractBuilder, allOpts: Required<ERC20Options>) {
     };
     c.addImplementedTrait(hooksTrait);
 
-    const beforeUpdateFn = c.addFunction(hooksTrait, {
-      name: 'before_update',
-      args: [
-        { name: 'ref self', type: 'ERC20Component::ComponentState<ContractState>' },
-        { name: 'from', type: 'ContractAddress' },
-        { name: 'recipient', type: 'ContractAddress' },
-        { name: 'amount', type: 'u256' },
-      ],
-      code: [],
-    });
-
-    const afterUpdateFn = c.addFunction(hooksTrait, {
-      name: 'after_update',
-      args: [
-        { name: 'ref self', type: 'ERC20Component::ComponentState<ContractState>' },
-        { name: 'from', type: 'ContractAddress' },
-        { name: 'recipient', type: 'ContractAddress' },
-        { name: 'amount', type: 'u256' },
-      ],
-      code: [],
-    });
-
     if (allOpts.pausable) {
+      const beforeUpdateFn = c.addFunction(hooksTrait, {
+        name: 'before_update',
+        args: [
+          { name: 'ref self', type: 'ERC20Component::ComponentState<ContractState>' },
+          { name: 'from', type: 'ContractAddress' },
+          { name: 'recipient', type: 'ContractAddress' },
+          { name: 'amount', type: 'u256' },
+        ],
+        code: [],
+      });
+
       beforeUpdateFn.code.push(
-        'let contract_state = ERC20Component::HasComponent::get_contract(@self);',
+        'let contract_state = self.get_contract();',
         'contract_state.pausable.assert_not_paused();',
       );
     }
@@ -148,9 +140,20 @@ function addHooks(c: ContractBuilder, allOpts: Required<ERC20Options>) {
 
       addVotesComponent(c, toFelt252(allOpts.appName, 'appName'), toFelt252(allOpts.appVersion, 'appVersion'));
 
+      const afterUpdateFn = c.addFunction(hooksTrait, {
+        name: 'after_update',
+        args: [
+          { name: 'ref self', type: 'ERC20Component::ComponentState<ContractState>' },
+          { name: 'from', type: 'ContractAddress' },
+          { name: 'recipient', type: 'ContractAddress' },
+          { name: 'amount', type: 'u256' },
+        ],
+        code: [],
+      });
+
       afterUpdateFn.code.push(
-        'let mut contract_state = ERC20Component::HasComponent::get_contract_mut(ref self);',
-        'contract_state.erc20_votes.transfer_voting_units(from, recipient, amount);',
+        'let mut contract_state = self.get_contract_mut();',
+        'contract_state.votes.transfer_voting_units(from, recipient, amount);',
       );
     }
   } else {
@@ -194,7 +197,7 @@ function addPremint(c: ContractBuilder, amount: string) {
 
     c.addStandaloneImport('starknet::ContractAddress');
     c.addConstructorArgument({ name:'recipient', type:'ContractAddress' });
-    c.addConstructorCode(`self.erc20.mint(recipient, ${premintAbsolute})`); 
+    c.addConstructorCode(`self.erc20.mint(recipient, ${premintAbsolute})`);
   }
 }
 
@@ -244,40 +247,6 @@ function addMintable(c: ContractBuilder, access: Access) {
   requireAccessControl(c, externalTrait, functions.mint, access, 'MINTER', 'minter');
 }
 
-function addVotesComponent(c: ContractBuilder, name: string, version: string) {
-  c.addComponent(components.ERC20VotesComponent, [], false);
-  c.addComponent(components.NoncesComponent, [], false);
-  c.addStandaloneImport('openzeppelin::token::erc20::extensions::ERC20VotesComponent::InternalTrait as ERC20VotesInternalTrait');
-  c.addStandaloneImport('openzeppelin::utils::cryptography::snip12::SNIP12Metadata');
-  c.addStandaloneImport('starknet::ContractAddress');
-
-  const SNIP12Metadata: BaseImplementedTrait = {
-    name: 'SNIP12MetadataImpl',
-    of: 'SNIP12Metadata',
-    tags: [],
-    priority: 0,
-  };
-  c.addImplementedTrait(SNIP12Metadata);
-
-  c.addFunction(SNIP12Metadata, {
-    name: 'name',
-    args: [],
-    returns: 'felt252',
-    code: [
-      `'${name}'`,
-    ],
-  });
-
-  c.addFunction(SNIP12Metadata, {
-    name: 'version',
-    args: [],
-    returns: 'felt252',
-    code: [
-      `'${version}'`,
-    ],
-  });
-}
-
 const components = defineComponents( {
   ERC20Component: {
     path: 'openzeppelin::token::erc20',
@@ -294,40 +263,6 @@ const components = defineComponents( {
       name: 'ERC20InternalImpl',
       value: 'ERC20Component::InternalImpl<ContractState>',
     },
-  },
-  ERC20VotesComponent: {
-    path: 'openzeppelin::token::erc20::extensions',
-    substorage: {
-      name: 'erc20_votes',
-      type: 'ERC20VotesComponent::Storage',
-    },
-    event: {
-      name: 'ERC20VotesEvent',
-      type: 'ERC20VotesComponent::Event',
-    },
-    impls: [
-      {
-        name: 'ERC20VotesComponentImpl',
-        value: 'ERC20VotesComponent::ERC20VotesImpl<ContractState>',
-      },
-    ],
-  },
-  NoncesComponent: {
-    path: 'openzeppelin::utils::cryptography::nonces',
-    substorage: {
-      name: 'nonces',
-      type: 'NoncesComponent::Storage',
-    },
-    event: {
-      name: 'NoncesEvent',
-      type: 'NoncesComponent::Event',
-    },
-    impls: [
-      {
-        name: 'NoncesImpl',
-        value: 'NoncesComponent::NoncesImpl<ContractState>',
-      },
-    ],
   },
 });