+++ packages/core/src/api.ts
@@ -2,6 +2,7 @@ import type { CommonOptions } from './common-options';
 import { printERC20, defaults as erc20defaults, isAccessControlRequired as erc20IsAccessControlRequired, ERC20Options } from './erc20';
 import { printERC721, defaults as erc721defaults, isAccessControlRequired as erc721IsAccessControlRequired, ERC721Options } from './erc721';
 import { printERC1155, defaults as erc1155defaults, isAccessControlRequired as erc1155IsAccessControlRequired, ERC1155Options } from './erc1155';
+import { printStablecoin, defaults as stablecoinDefaults, isAccessControlRequired as stablecoinIsAccessControlRequired, StablecoinOptions } from './stablecoin';
 import { printGovernor, defaults as governorDefaults, isAccessControlRequired as governorIsAccessControlRequired, GovernorOptions } from './governor';
 import { printCustom, defaults as customDefaults, isAccessControlRequired as customIsAccessControlRequired, CustomOptions } from './custom';
 
@@ -26,6 +27,8 @@ export interface WizardContractAPI<Options extends CommonOptions> {
 export type ERC20 = WizardContractAPI<ERC20Options>;
 export type ERC721 = WizardContractAPI<ERC721Options>;
 export type ERC1155 = WizardContractAPI<ERC1155Options>;
+export type Stablecoin = WizardContractAPI<StablecoinOptions>;
+export type RealWorldAsset = WizardContractAPI<StablecoinOptions>;
 export type Governor = WizardContractAPI<GovernorOptions>;
 export type Custom = WizardContractAPI<CustomOptions>;
 
@@ -43,6 +46,16 @@ export const erc1155: ERC1155 = {
   print: printERC1155,
   defaults: erc1155defaults,
   isAccessControlRequired: erc1155IsAccessControlRequired
+}
+export const stablecoin: Stablecoin = {
+  print: printStablecoin,
+  defaults: stablecoinDefaults,
+  isAccessControlRequired: stablecoinIsAccessControlRequired
+}
+export const realWorldAsset: RealWorldAsset = {
+  print: printStablecoin,
+  defaults: stablecoinDefaults,
+  isAccessControlRequired: stablecoinIsAccessControlRequired
 }
 export const governor: Governor = {
   print: printGovernor,