+++ packages/core/src/build-generic.ts
@@ -2,12 +2,15 @@ import { CustomOptions, buildCustom } from './custom';
 import { ERC20Options, buildERC20 } from './erc20';
 import { ERC721Options, buildERC721 } from './erc721';
 import { ERC1155Options, buildERC1155 } from './erc1155';
+import { StablecoinOptions, buildStablecoin } from './stablecoin';
 import { GovernorOptions, buildGovernor } from './governor';
 
 export interface KindedOptions {
   ERC20:    { kind: 'ERC20' }    & ERC20Options;
   ERC721:   { kind: 'ERC721' }   & ERC721Options;
   ERC1155:  { kind: 'ERC1155' }  & ERC1155Options;
+  Stablecoin:  { kind: 'Stablecoin' }  & StablecoinOptions;
+  RealWorldAsset:  { kind: 'RealWorldAsset' }  & StablecoinOptions;
   Governor: { kind: 'Governor' } & GovernorOptions;
   Custom:  { kind: 'Custom' }  & CustomOptions;
 }
@@ -24,6 +27,12 @@ export function buildGeneric(opts: GenericOptions) {
 
     case 'ERC1155':
       return buildERC1155(opts);
+
+    case 'Stablecoin':
+      return buildStablecoin(opts);
+
+    case 'RealWorldAsset':
+      return buildStablecoin(opts);
 
     case 'Governor':
       return buildGovernor(opts);